ToUnixTime fix
I changed the ToUnixTime helper method to return the total milliseconds since epoch instead of the milliseconds of the DateTime object. Before this, if the user provided the optional Timestamp in a Point object, the data would be written to InfluxDB as having happened a few milliseconds after the start of epoch.